requirements.txt
Flask==2.3.2
gevent
python-magic
logfire
